[Test] move methods from bwc test to test package for use in other plugins
